+++ src/main/java/com/kz/internship_project/service/impl/CourseServiceImpl.java
@@ -4,13 +4,21 @@ import com.kz.internship_project.dto.CourseCreateDto;
 import com.kz.internship_project.dto.CourseResponseDto;
 import com.kz.internship_project.entity.Course;
 import com.kz.internship_project.mapper.CourseMapper;
+import com.kz.internship_project.repository.ChapterRepository;
 import com.kz.internship_project.repository.CourseRepository;
 import com.kz.internship_project.service.CourseService;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Slf4j
 @Service
@@ -20,6 +28,7 @@ public class CourseServiceImpl implements CourseService {
 
     private final CourseRepository courseRepository;
     private final CourseMapper courseMapper;
+    private final ChapterRepository chapterRepository;
 
     @Override
     @Transactional
@@ -69,6 +78,24 @@ public class CourseServiceImpl implements CourseService {
         if (!courseRepository.existsById(id)) {
             throw new EntityNotFoundException("Курс с id " + id + " не найден");
         }
+
+        if (chapterRepository.existsById(id)){
+            throw new IllegalArgumentException("Курс нельзя удалить, пока в нем есть главы!");
+        }
         courseRepository.deleteById(id);
     }
+
+    @Override
+    public Page<CourseResponseDto> getAll(int page, int size, String sortBy, String sortDir){
+
+        Sort sort = sortDir.equalsIgnoreCase(Sort.Direction.DESC.name())
+                ? Sort.by(sortBy).descending()
+                : Sort.by(sortBy).ascending();
+
+        Pageable pageable = PageRequest.of(page, size, sort);
+        Page<Course> courses =courseRepository.findAll(pageable);
+
+        return courses.map(courseMapper::toDto);
+    }
+
 }
